+++ docs/.vitepress/theme/index.ts
@@ -1,0 +1,20 @@
+// https://vitepress.dev/guide/custom-theme
+import { h } from 'vue'
+import type { EnhanceAppContext } from 'vitepress'
+import DefaultTheme from 'vitepress/theme'
+import TwoSlashFloatingVue from 'vitepress-plugin-twoslash/client'
+import CustomLayout from './CustomLayout.vue'
+
+import 'vitepress-plugin-twoslash/style.css'
+import 'uno.css'
+import './style.css'
+
+export default {
+  extends: DefaultTheme,
+  Layout: () => {
+    return h(CustomLayout)
+  },
+  enhanceApp({ app }: EnhanceAppContext) {
+    app.use(TwoSlashFloatingVue as any)
+  },
+}