+++ src/define.ts
@@ -1,5 +1,6 @@
 import { getEsbuildPlugin } from './esbuild'
 import { getFarmPlugin } from './farm'
+import { getRolldownPlugin } from './rolldown'
 import { getRollupPlugin } from './rollup'
 import { getRspackPlugin } from './rspack'
 import type { UnpluginFactory, UnpluginInstance } from './types'
@@ -18,6 +19,10 @@ export function createUnplugin<UserOptions, Nested extends boolean = boolean>(
     },
     get vite() {
       return getVitePlugin(factory)
+    },
+    /** @experimental do not use it in production */
+    get rolldown() {
+      return getRolldownPlugin(factory)
     },
     get webpack() {
       return getWebpackPlugin(factory)
@@ -53,12 +58,20 @@ export function createVitePlugin<UserOptions, Nested extends boolean = boolean>(
   return getVitePlugin(factory)
 }
 
+/** @experimental do not use it in production */
+export function createRolldownPlugin<UserOptions, Nested extends boolean = boolean>(
+  factory: UnpluginFactory<UserOptions, Nested>,
+) {
+  return getRolldownPlugin(factory)
+}
+
 export function createWebpackPlugin<UserOptions, Nested extends boolean = boolean>(
   factory: UnpluginFactory<UserOptions, Nested>,
 ) {
   return getWebpackPlugin(factory)
 }
 
+/** @experimental do not use it in production */
 export function createRspackPlugin<UserOptions, Nested extends boolean = boolean>(
   factory: UnpluginFactory<UserOptions, Nested>,
 ) {