+++ src/rspack/loaders/transform.ts
@@ -1,6 +1,5 @@
 import type { LoaderContext } from '@rspack/core'
-import { createRspackContext } from '../context'
-import type { UnpluginContext } from '../../types'
+import { createBuildContext, createContext } from '../context'
 
 export default async function transform(
   this: LoaderContext,
@@ -24,15 +23,10 @@ export default async function transform(
   if (!plugin?.transform)
     return callback(null, source, map)
 
-  const context: UnpluginContext = {
-    error: error =>
-      this.emitError(typeof error === 'string' ? new Error(error) : error),
-    warn: error =>
-      this.emitWarning(typeof error === 'string' ? new Error(error) : error),
-  }
+  const context = createContext(this)
   const res = await plugin.transform.call(
     Object.assign(
-      this._compilation && createRspackContext(this._compilation),
+      this._compilation && createBuildContext(this._compilation),
       context,
     ),
     source,