+++ src/webpack/loaders/transform.ts
@@ -1,6 +1,5 @@
 import type { LoaderContext } from 'webpack'
-import type { UnpluginContext } from '../../types'
-import { createContext } from '../context'
+import { createBuildContext, createContext } from '../context'
 
 export default async function transform(this: LoaderContext<{ unpluginName: string }>, source: string, map: any) {
   const callback = this.async()
@@ -19,12 +18,16 @@ export default async function transform(this: LoaderContext<{ unpluginName: stri
   if (!plugin?.transform)
     return callback(null, source, map)
 
-  const context: UnpluginContext = {
-    error: error => this.emitError(typeof error === 'string' ? new Error(error) : error),
-    warn: error => this.emitWarning(typeof error === 'string' ? new Error(error) : error),
-  }
+  const context = createContext(this)
   const res = await plugin.transform.call(
-    { ...this._compilation && createContext(this._compilation) as any, ...context },
+    { ...createBuildContext({
+      addWatchFile: (file) => {
+        this.addDependency(file)
+      },
+      getWatchFiles: () => {
+        return this.getDependencies()
+      },
+    }, this._compilation), ...context },
     source,
     this.resource,
   )